Adding missing serf commands and cluster CLI

+++ kong/cmd/utils/serf_signals.lua
@@ -10,10 +10,12 @@ local pl_path = require "pl.path"
 local pl_file = require "pl.file"
 local kill = require "kong.cmd.utils.kill"
 local log = require "kong.cmd.utils.log"
+local utils = require "kong.tools.utils"
 local fmt = string.format
 
 local serf_bin_name = "serf"
 local serf_pid_name = "serf.pid"
+local serf_node_id = "serf.id"
 local serf_event_name = "kong"
 local start_timeout = 2
 
@@ -55,11 +57,27 @@ client:request { \
 resty -e "$CMD"
 ]]
 
+local function prepare_identifier(kong_config, nginx_prefix)
+  local id_path = pl_path.join(nginx_prefix, serf_node_id)
+  if not pl_path.exists(id_path) then
+    local id = utils.get_hostname().."_"..kong_config.cluster_listen.."_"..utils.random_string()
+
+    log.verbose("saving Serf identifier in %s", id_path)
+    local ok, err = pl_file.write(id_path, id)
+    if not ok then return nil, err end
+  end
+  return true
+end
+
 local function prepare_prefix(kong_config, nginx_prefix, script_path)
+  local ok, err = prepare_identifier(kong_config, nginx_prefix)
+  if not ok then return nil, err end
+
   log.verbose("dumping Serf shell script handler in %s", script_path)
   local script = fmt(script_template, "127.0.0.1", kong_config.admin_port)
 
-  pl_file.write(script_path, script)
+  local ok, err = pl_file.write(script_path, script)
+  if not ok then return nil, err end
   local ok, _, _, stderr = pl_utils.executeex("chmod +x "..script_path)
   if not ok then return nil, stderr end
 
@@ -85,19 +103,19 @@ function _M.start(kong_config, nginx_prefix, dao)
     pl_file.delete(pid_path)
   end
 
+  -- prepare shell script
+  local script_path = pl_path.join(nginx_prefix, "serf_event.sh")
+  local ok, err = prepare_prefix(kong_config, nginx_prefix, script_path)
+  if not ok then return nil, err end
+
   -- make sure Serf is in PATH
   local ok, err = check_serf_bin()
   if not ok then return nil, err end
 
-  local serf = Serf.new(kong_config, dao)
+  local serf = Serf.new(kong_config, nginx_prefix, dao)
 
   local node_name = serf.node_name
-  local script_path = pl_path.join(nginx_prefix, "serf_event.sh")
   local log_path = pl_path.join(nginx_prefix, "serf.log")
-
-  -- prepare shell script
-  local ok, err = prepare_prefix(kong_config, nginx_prefix, script_path)
-  if not ok then return nil, err end
 
   local args = setmetatable({
     ["-bind"] = kong_config.cluster_listen,
@@ -152,7 +170,14 @@ function _M.start(kong_config, nginx_prefix, dao)
   return true
 end
 
-function _M.stop(nginx_prefix)
+function _M.stop(kong_config, nginx_prefix, dao)
+  log.info("Leaving cluster")
+
+  local serf = Serf.new(kong_config, nginx_prefix, dao)
+
+  local ok, err = serf:leave()
+  if not ok then return nil, err end
+
   local pid_path = pl_path.join(nginx_prefix, serf_pid_name)
   log.verbose("stopping Serf agent at %s", pid_path)
   return kill(pid_path, "-9")
